refactor: Move IPluginAuth interface to types directory for better organization and update pluginAuth schema to use the new import
